Provide separate auto-polyfill version of this module
This commit provides a mechanism for users to load a version of this
module which automatically installs the polyfill, separately from the
default non-automatic version, as outlined in
https://github.com/jakearchibald/es6-promise/pull/126#issuecomment-120542090

It uses require('es6-promise/auto') instead of
require('es6-promise/polyfil') so that it does not directly conflict
with the es-shim-api API Contract[1], in case support is later added.

In addition to the Node/Browserify file, a UMD file which triggers
automatic polyfill installation is also provided.

Fixes: #140
Alternative-to: #126

1.  https://github.com/es-shims/es-shim-api#api-contract

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>